refactor: Reworked context menu for CategoryListView
Context menu management of the CategoryListView has a
leaky encapsulation with the menu being managed and
set by controls other than CategoryListView.

Now the context menu is exclusively managed by
CategoryListView. The context menu setup has been
split into two distinct groups:
- category, and
- mod.
This allows for cleaner maintenance and the menu is
tailor-made for the type of item(s) being managed.
There are provisions for multi-item management too.

The context menu has received a number of new items,
which whilst been available to the user, were not
very obvious to the user.

These include (not limited to):
- Category -> Expand/collapse all
- Category -> Mod update warning -> ...
- Category -> Show/hide empty
- Mod -> Activate/deactivate
- Mod -> Uninstall -> current profile/all profiles/
permanent delete
- Mod -> Mod update warning -> ...

Several menus have been reworked to provide cleaner UI.
For example "Mod open readme" - readme files are scanned
for during the context menu opening and renders the menu
depending on presence of the files.

All menus are context-aware, i.e. they are conditionally
rendered depending on the state of the selected item(s)
or global settings.