Fixed path to registry location for HKLM - whole path was being provided which was incorrect.
CreateSubKey method was throwing an IO exception because the key was previously opened. CreateSubKey can actually Open OR Create a key on the fly, the double access is not allowed.

The p_strHDLink parameter is null off the bat causing a NullReferenceException when attempting to save a value to the registry. Checking for null to stop this from killing the whole process.