Auto creating missing folders, because you are going to need them no matter what. Auto creating the folders for Script Types and Game Modes.
If Game Modes are not found throw an exception, if running in debug add a clue as to what is wrong. I did this to stop a DivideByZeroError from occurring in the GameDectionForm.ViewModel property.

If Script Types are not and running in debug add a clue as to what is wrong. It wasn't obvious what the problem was.

GameDetectionForm.ViewModel - cleaned this property up to make it more legible. Consider using a method, this is doing too much, additionally it can throw a DivideByZeroException.