Fixed SecurityException in C# install scripts
Path.GetTempPath() was called in the CSharpScript context due to m_strTempPath being null in the EnvironmentInfo (not filled in the settings file). This caused a SecurityException and failure to install any files from C# install scripts due to the context lacking the permission to read environment variables.

m_strTempPath is now filled when the EnvironmentInfo is created instead of calling Path.GetTempPath() in the TemporaryPath getter.