Txt file overwrite overridden
The program will no longer ask you whether to overwrite .txt files (it just won't do that), it turns out it's 99.9% of the times just a false positive for ReadMe and other useless text files.
Should you find any occurrence where this is needed, then let me know and I will add a Settings options to turn this on and off.